Adding support for metadata files.
metadata files should mimic the file name, adding the extesion
'.metadata' to the end of the file. For instance, 'a.pdf' and
'a.pdf.metadata'.

The metadata file should be a key/value pair with the same structure of
a .properties file. The file should also be created using utf-8.

Each pair will be added to the lucene document of that file, being the
key the name of the field and value the stored information. Besides
being stored, that value will be analyzed, thus searchable.